fix: integration tests

+++ internal/pkg/api/repository/testresult/mongo_test.go
@@ -94,13 +94,13 @@ func TestTestExecutionsMetrics(t *testing.T) {
 	t.Run("limit should limit executions", func(t *testing.T) {
 		metrics, err := repository.GetTestSuiteMetrics(context.Background(), testName, 1, 100)
 		assert.NoError(err)
-		assert.Equal(1, metrics.TotalExecutions)
+		assert.Equal(int32(1), metrics.TotalExecutions)
 	})
 
 	t.Run("filter last n days should limit executions", func(t *testing.T) {
 		metrics, err := repository.GetTestSuiteMetrics(context.Background(), testName, 100, 1)
 		assert.NoError(err)
-		assert.Equal(19, metrics.TotalExecutions)
+		assert.Equal(int32(19), metrics.TotalExecutions)
 	})
 }
